requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.60.0
matplotlib>=3.4.0
pyyaml>=6.0
pytest>=7.0.0
